refactor: extract workspace emergency cleanup
